Updated zotero and lib and styles
I was getting warnings because it couldn't find
certain references because the library had not
been exported recently so I exported a new version.
Also, it was throwing an error about not finding
the ecology style even though I had been using
the ecological monographs style instead.
To shut it up, I just downloaded the ecology style
and put it in the zotero styles directory of this
project.